pyproject.toml
[project]
name = "strava-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
